+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: maratojo <maratojo@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/16 08:27:15 by maratojo          #+#    #+#             */
-/*   Updated: 2026/02/16 09:00:36 by maratojo         ###   ########.fr       */
+/*   Updated: 2026/02/17 10:39:05 by maratojo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_printf(const char *format, ...);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
 int	ft_putptr(void *ptr);
-int	ft_puthexa(unsigned long n, char *base);
+int	ft_puthexa(unsigned int n, char *base);
 int	ft_putnbr(int n);
 
 #endif
